Add The Insertion Sort Algorithm.
Insertion sort works similarly as we sort cards in our hand in a card
game. Here, a sub-list is maintained which is always sorted. An
element which is to be 'insert'ed  in this sorted sub-list, has to find
its appropriate place, and then it has to be inserted there. This is
better than selection sort for a small list.

Insertion Sort:
Worst Case Time Complexity: O(n^2)
Best Case Time Complexity: O(n)
Average Case Time Complexity: O(n^2)
Space complexity: O(1)