Add User-Friendly Elements.
Add a class called CoutColor.
This will add functionality to change the colors in osstream.
This functionality is used to highlight the swapping of two
numbers during a sorting process.

Add a class called SimpleVisualizer (Not completed yet).
This class will use a pre-defined integer array with 10
elements and visualize how algorithms work. For each step,
it will draw the array and highlight the process of swapping
numbers.

ToDO: Currently only works for BubbleSort. Yet to implement
support for other algorithms.

Add a class called UserUI.
Separate out the UI elements from the main to keep it clean.
UserUI has 3 modes.
Visualize: Draw array and highlight the swapping.
Sort: Let user-defined an array with numbers and sort it.
Graph: Draw a graph and provide all the analytic.
(Not implemented yet.)

Add new modules to Make and Scons.